% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_osm_power_data.R
\name{get_tidy_osm_power_data}
\alias{get_tidy_osm_power_data}
\title{Get tidy osm power data}
\usage{
get_tidy_osm_power_data(region)
}
\arguments{
\item{region}{}
}
\value{

}
\description{
Get tidy osm power data
}
\examples{
get_tidy_osm_power_data("Zuid-Holland")
}
